speed increases
Some of the files may be included in this commit simply because of new
line endings. I worked on this in one of my classes on a computer that
didn't initially have the source, so I had to get it from GitHub which
had Unix line endings whereas Visual Studio likes Windows line endings.